update classes and interfaces names
