+++ Common/Common.csproj
@@ -32,19 +32,14 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IDistributedTask.cs" />
     <Compile Include="ITask.cs" />
+    <Compile Include="ISubTask.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ITaskFactory.cs" />
     <Compile Include="TaskCreator.cs" />
+    <Compile Include="ISubtaskFactory.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
