+++ Common/ISubTask.cs
@@ -1,0 +1,8 @@
+﻿namespace Common
+{
+    public interface ISubTask
+    {
+        // TODO: use more developer-friendly types instead of string
+        string Perform(string input);
+    }
+}
