+++ Common/ISubTaskFactory.cs
@@ -1,12 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace Common
 {
-    public interface ITaskFactory
+    public interface ISubTaskFactory
     {
         // TODO: use more developer-friendly types instead of string
         void CreateNewTask(string input);
