+++ Common/ITask.cs
@@ -1,8 +1,16 @@
-﻿namespace Common
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Common
 {
     public interface ITask
     {
         // TODO: use more developer-friendly types instead of string
-        string Perform(string input);
+        // TODO: reference corresponding ISubTask from ITask
+        void DefineTasks(string input, ISubTaskFactory subTaskFactory);
+
+        string AggregateResults(string input, string[] results);
     }
 }
