+++ Inzynierka.sln
@@ -5,11 +5,9 @@ VisualStudioVersion = 15.0.27703.2035
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "Common\Common.csproj", "{BB8E5B5E-D391-4B1D-BF6E-C2997504B57C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FactorialSumDistributed", "FactorialSumDistributed\FactorialSumDistributed.csproj", "{2E674339-2E68-4E8F-A5A7-00FF2DBBE1D1}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FactorialTask", "FactorialTask\FactorialTask.csproj", "{5F3654E6-ED74-4D7A-BFD5-AB32540728E3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServerRunnerCore", "ServerRunner\ServerRunner.csproj", "{CEFF0D92-8935-4D85-87E9-6B01022EE10D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServerRunner", "ServerRunner\ServerRunner.csproj", "{CEFF0D92-8935-4D85-87E9-6B01022EE10D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,14 +19,6 @@ Global
 		{BB8E5B5E-D391-4B1D-BF6E-C2997504B57C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB8E5B5E-D391-4B1D-BF6E-C2997504B57C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB8E5B5E-D391-4B1D-BF6E-C2997504B57C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2E674339-2E68-4E8F-A5A7-00FF2DBBE1D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E674339-2E68-4E8F-A5A7-00FF2DBBE1D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E674339-2E68-4E8F-A5A7-00FF2DBBE1D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E674339-2E68-4E8F-A5A7-00FF2DBBE1D1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4DD0D522-7738-4F96-AE7B-DBE7F76FBFF4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4DD0D522-7738-4F96-AE7B-DBE7F76FBFF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4DD0D522-7738-4F96-AE7B-DBE7F76FBFF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4DD0D522-7738-4F96-AE7B-DBE7F76FBFF4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5F3654E6-ED74-4D7A-BFD5-AB32540728E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5F3654E6-ED74-4D7A-BFD5-AB32540728E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5F3654E6-ED74-4D7A-BFD5-AB32540728E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
